Driver can now advance to next genre
Driver now calls populate_state() on init. This loads the list of genres
in from the page (which is scraped). Not smart enough to tell which
genre it started on based on the url.

Had to add a method to Scraper to return the BeautifulSoup tags, instead
of just a url list.