Driver: current letter and page now use Tags
Changed the way Driver sets initial state and handles the current letter
and current page. Instead of maintaining a single letter and an integer
(which weren't being used for Driver.next_url() anyway), use the actual
Tags with urls. This obsoleted the old Driver.next_letter() method.

Also found an edge case (or maybe it was already there) where
Driver.write_urls_to_file() can't use driver.parse_url() to populate
current_letter and current_page. This is because the url for the "#"
page is "letter=*".