Moved Driver.populate_state out of init
Moved Driver.populate_state() out of __init__(). Since we want to scrape the
first page also when we do our crawl, it doesn't really belong there,
since it fetches the first page. Also, it seems odd to have something
that does an http request inside the constructor of a class.